* Source/GSDragView.m (-init): Set the window background colour to clear. * Headers/AppKit/NSAffineTransform.h, * Headers/AppKit/NSView.h, * Source/NSAffineTransform.m, * Source/NSPrintOperation.m, * Source/NSView.m: Use CGFlaot instead of float. * Tests/gui/NSView/NSView_bounds_scale.m: New test file that currently fails.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@33398 72102866-910b-0410-8b05-ffd578937521

+++ Headers/AppKit/NSAffineTransform.h
@@ -45,11 +45,11 @@
 #if GS_API_VERSION(GS_API_NONE, 011500)
 @interface NSAffineTransform (GNUstep)
 - (void) translateToPoint: (NSPoint)point;
-- (void) scaleTo: (float)sx : (float)sy;
+- (void) scaleTo: (CGFloat)sx : (CGFloat)sy;
 - (void) makeIdentityMatrix;
-- (float) rotationAngle;
+- (CGFloat) rotationAngle;
 - (void) setFrameOrigin: (NSPoint)point;
-- (void) setFrameRotation: (float)angle;
+- (void) setFrameRotation: (CGFloat)angle;
 
 /* Deprecated: use -invert  */
 - (void) inverse;
